Se agrega el ABM de perfiles por usuario

+++ src/main/java/py/com/domainsoft/seguridad/repositories/UsuarioRepository.java
@@ -8,6 +8,6 @@ import py.com.domainsoft.seguridad.entities.UsuarioEntity;
  * @author mcespedes
  */
 public interface UsuarioRepository
-        extends PagingAndSortingRepository<UsuarioEntity, Long>  {
+        extends PagingAndSortingRepository<UsuarioEntity, Integer>  {
     UsuarioEntity findByLogin(String login);
 }
